Test.Cardano.Crypto.Gen

Protocol Magic Generator

genProtocolMagic ∷ Gen ProtocolMagic

genProtocolMagicId ∷ Gen ProtocolMagicId

Whilst ProtocolMagicId is represented as a Word32 in cardano-ledger, in cardano-sl it was an Int32. In order to tolerate this, and since we don't care about testing compatibility with negative values, we only generate values between 0 and (maxBound :: Int32) - 1, inclusive.

genRequiresNetworkMagic ∷ Gen RequiresNetworkMagic

Sign Tag Generator

genSignTag ∷ Gen SignTag

Key Generators

genKeypair ∷ Gen (VerificationKey, SigningKey)

genVerificationKey ∷ Gen VerificationKey

genSigningKey ∷ Gen SigningKey

Redeem Key Generators

genRedeemKeypair ∷ Gen (RedeemVerificationKey, RedeemSigningKey)

genRedeemVerificationKey ∷ Gen RedeemVerificationKey

genCompactRedeemVerificationKey ∷ Gen CompactRedeemVerificationKey

genRedeemSigningKey ∷ Gen RedeemSigningKey

Signature Generators

genSignature ∷ ToCBOR a ⇒ ProtocolMagicId → Gen a → Gen (Signature a)

genSignatureEncoded ∷ Gen ByteString → Gen (Signature a)

genRedeemSignature ∷ ToCBOR a ⇒ ProtocolMagicId → Gen a → Gen (RedeemSignature a)

Hash Generators

genAbstractHash ∷ (ToCBOR a, HashAlgorithm algo) ⇒ Gen a → Gen (AbstractHash algo a)

SafeSigner Generators

genSafeSigner ∷ Gen SafeSigner

PassPhrase Generators

genPassPhrase ∷ Gen PassPhrase

Helper Generators

genHashRaw ∷ Gen (Hash Raw)

genTextHash ∷ Gen (Hash Text)

feedPM ∷ (ProtocolMagicId → Gen a) → Gen a